Implement location detection

+++ app/src/main/java/edu/pk/weatherapp/model/Wind.kt
@@ -3,4 +3,4 @@ package edu.pk.weatherapp.model
 import com.squareup.moshi.Json
 
 data class Wind(@Json(name = "speed") var speed: Double,
-                @Json(name = "deg") var deg: Int = 0)
+                @Json(name = "deg") var deg: Double = 0.0)
